+++ proxmox/virtual_environment_users.go
@@ -27,9 +27,30 @@ func (c *VirtualEnvironmentClient) CreateUser(d *VirtualEnvironmentUserCreateReq
 	return c.DoRequest(hmPOST, "access/users", d, nil)
 }
 
+// CreateUserToken creates a user token.
+func (c *VirtualEnvironmentClient) CreateUserToken(id, tokenID string, d *VirtualEnvironmentUserTokenCreateRequestBody) (*VirtualEnvironmentUserTokenCreateResponseData, error) {
+	resBody := &VirtualEnvironmentUserTokenCreateResponseBody{}
+	err := c.DoRequest(hmPOST, fmt.Sprintf("access/users/%s/token/%s", url.PathEscape(id), url.PathEscape(tokenID)), d, resBody)
+
+	if err != nil {
+		return nil, err
+	}
+
+	if resBody.Data == nil {
+		return nil, errors.New("The server did not include a data object in the response")
+	}
+
+	return resBody.Data, nil
+}
+
 // DeleteUser deletes an user.
 func (c *VirtualEnvironmentClient) DeleteUser(id string) error {
 	return c.DoRequest(hmDELETE, fmt.Sprintf("access/users/%s", url.PathEscape(id)), nil, nil)
+}
+
+// DeleteUserToken deletes a user token.
+func (c *VirtualEnvironmentClient) DeleteUserToken(id, tokenID string) error {
+	return c.DoRequest(hmDELETE, fmt.Sprintf("access/users/%s/token/%s", url.PathEscape(id), url.PathEscape(tokenID)), nil, nil)
 }
 
 // GetUser retrieves an user.
@@ -52,6 +73,31 @@ func (c *VirtualEnvironmentClient) GetUser(id string) (*VirtualEnvironmentUserGe
 
 	if resBody.Data.Groups != nil {
 		sort.Strings(*resBody.Data.Groups)
+	}
+
+	if resBody.Data.Tokens != nil {
+		for _, tv := range *resBody.Data.Tokens {
+			if tv.ExpirationDate != nil {
+				expirationDate := CustomTimestamp(time.Time(*tv.ExpirationDate).UTC())
+				tv.ExpirationDate = &expirationDate
+			}
+		}
+	}
+
+	return resBody.Data, nil
+}
+
+// GetUserToken retrieves a user token.
+func (c *VirtualEnvironmentClient) GetUserToken(id, tokenID string) (*VirtualEnvironmentUserTokenGetResponseData, error) {
+	resBody := &VirtualEnvironmentUserTokenGetResponseBody{}
+	err := c.DoRequest(hmGET, fmt.Sprintf("access/users/%s/token/%s", url.PathEscape(id), url.PathEscape(tokenID)), nil, resBody)
+
+	if err != nil {
+		return nil, err
+	}
+
+	if resBody.Data == nil {
+		return nil, errors.New("The server did not include a data object in the response")
 	}
 
 	return resBody.Data, nil
@@ -82,14 +128,46 @@ func (c *VirtualEnvironmentClient) ListUsers(enabled, full bool) ([]*VirtualEnvi
 		return resBody.Data[i].ID < resBody.Data[j].ID
 	})
 
-	for i := range resBody.Data {
-		if resBody.Data[i].ExpirationDate != nil {
-			expirationDate := CustomTimestamp(time.Time(*resBody.Data[i].ExpirationDate).UTC())
-			resBody.Data[i].ExpirationDate = &expirationDate
+	for _, v := range resBody.Data {
+		if v.ExpirationDate != nil {
+			expirationDate := CustomTimestamp(time.Time(*v.ExpirationDate).UTC())
+			v.ExpirationDate = &expirationDate
 		}
 
-		if resBody.Data[i].Groups != nil {
-			sort.Strings(*resBody.Data[i].Groups)
+		if v.Groups != nil {
+			sort.Strings(*v.Groups)
+		}
+
+		if v.Tokens != nil {
+			for _, tv := range *v.Tokens {
+				if tv.ExpirationDate != nil {
+					expirationDate := CustomTimestamp(time.Time(*tv.ExpirationDate).UTC())
+					tv.ExpirationDate = &expirationDate
+				}
+			}
+		}
+	}
+
+	return resBody.Data, nil
+}
+
+// ListUserTokens retrieves information about API tokens for a user.
+func (c *VirtualEnvironmentClient) ListUserTokens(id string) ([]*VirtualEnvironmentUserTokenListResponseData, error) {
+	resBody := &VirtualEnvironmentUserTokenListResponseBody{}
+	err := c.DoRequest(hmGET, fmt.Sprintf("access/users/%s/token", url.PathEscape(id)), nil, resBody)
+
+	if err != nil {
+		return nil, err
+	}
+
+	if resBody.Data == nil {
+		return nil, errors.New("The server did not include a data object in the response")
+	}
+
+	for _, tv := range resBody.Data {
+		if tv.ExpirationDate != nil {
+			expirationDate := CustomTimestamp(time.Time(*tv.ExpirationDate).UTC())
+			tv.ExpirationDate = &expirationDate
 		}
 	}
 
@@ -99,4 +177,20 @@ func (c *VirtualEnvironmentClient) ListUsers(enabled, full bool) ([]*VirtualEnvi
 // UpdateUser updates an user.
 func (c *VirtualEnvironmentClient) UpdateUser(id string, d *VirtualEnvironmentUserUpdateRequestBody) error {
 	return c.DoRequest(hmPUT, fmt.Sprintf("access/users/%s", url.PathEscape(id)), d, nil)
+}
+
+// UpdateUserToken updates a user token.
+func (c *VirtualEnvironmentClient) UpdateUserToken(id, tokenID string, d *VirtualEnvironmentUserTokenUpdateRequestBody) (*VirtualEnvironmentUserTokenUpdateResponseData, error) {
+	resBody := &VirtualEnvironmentUserTokenUpdateResponseBody{}
+	err := c.DoRequest(hmPUT, fmt.Sprintf("access/users/%s/token/%s", url.PathEscape(id), url.PathEscape(tokenID)), d, resBody)
+
+	if err != nil {
+		return nil, err
+	}
+
+	if resBody.Data == nil {
+		return nil, errors.New("The server did not include a data object in the response")
+	}
+
+	return resBody.Data, nil
 }