+++ proxmoxtf/provider.go
@@ -71,6 +71,7 @@ func Provider() *schema.Provider {
 			"proxmox_virtual_environment_role":          resourceVirtualEnvironmentRole(),
 			"proxmox_virtual_environment_time":          resourceVirtualEnvironmentTime(),
 			"proxmox_virtual_environment_user":          resourceVirtualEnvironmentUser(),
+			"proxmox_virtual_environment_user_token":    resourceVirtualEnvironmentUserToken(),
 			"proxmox_virtual_environment_vm":            resourceVirtualEnvironmentVM(),
 		},
 		Schema: map[string]*schema.Schema{